Move to compat module to avoid namespace collisions in salt.utils
