+++ salt/modules/grains.py
@@ -15,7 +15,7 @@ import collections
 from functools import reduce
 
 # Import 3rd-party libs
-import salt.utils.copy
+import salt.utils.compat
 from salt.utils.odict import OrderedDict
 import yaml
 import salt.ext.six as six
@@ -246,8 +246,8 @@ def setvals(grains, destructive=False):
         # This likely means we are running under Python 2.6 which cannot deepcopy
         # bound methods. Fallback to a modification of deepcopy which can support
         # this behavoir.
-        yaml_reps = salt.utils.copy.deepcopy_bound(yaml.representer.SafeRepresenter.yaml_representers)
-        yaml_multi_reps = salt.utils.copy.deepcopy_bound(yaml.representer.SafeRepresenter.yaml_multi_representers)
+        yaml_reps = salt.utils.compat.deepcopy_bound(yaml.representer.SafeRepresenter.yaml_representers)
+        yaml_multi_reps = salt.utils.compat.deepcopy_bound(yaml.representer.SafeRepresenter.yaml_multi_representers)
     yaml.representer.SafeRepresenter.add_representer(collections.defaultdict,
             yaml.representer.SafeRepresenter.represent_dict)
     yaml.representer.SafeRepresenter.add_representer(OrderedDict,
